Update tests to use the RomeoAPIStub to fetch journals
